feat(markdown-converter): add carbon classes option + fixes
